# Conflicts:
#	FinalProject/.classpath
#	FinalProject/.settings/org.eclipse.wst.common.component
#	FinalProject/bin/target/m2e-wtp/web-resources/META-INF/maven/com.team12/main/pom.properties
#	FinalProject/src/main/java/com/team12/main/team1/shop/review/Team1ReviewDAO.java
#	FinalProject/src/main/java/com/team12/main/team1/shop/review/Team1ReviewMapper.java
#	FinalProject/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml

+++ FinalProject/src/main/java/com/team12/main/team1/shop/review/Team1ReviewMapper.java
@@ -21,12 +21,12 @@ public interface Team1ReviewMapper {
 	int getAllReview(Team1ReviewDTO review);
 	
 	// 전체 별점 평균 구하기
-
-	int getAllReviewAvg();
+	double getAllReviewAvg();
 
 	
 	// 리뷰 수정하기
 	String updateReview(Team1ReviewDTO review);
+
